+++ src/stuff.h
@@ -13,7 +13,7 @@
 #include "global.h"
 
 /* Prototypes begin here */
-ptrdiff_t my_log2(size_t x);
+int my_log2(size_t x);
 int count_bits(unsigned INT32 x);
 int is_more_than_one_bit(unsigned INT32 x);
 double my_strtod(char *nptr, char **endptr);
